pyproject.toml
[tool.poetry]
name = "auth"
version = "0.1.0"
description = "auth micro service"
authors = ["嘉然今天吃什么 <jsclndnz@outlook.com>"]
license = "Apache"

[tool.poetry.dependencies]
python = "^3.9"
tortoise-orm = "^0.18.1"
aiomysql = "^0.0.22"
aiosqlite = "^0.17.0"
pycryptodome = "^3.14.1"
aerich = "^0.6.2"
pydantic = "^1.9.0"
PyJWT = { extras = ["crypto"], version = "^2.3.0" }
email-validator = "^1.1.3"
aiocache = { extras = ["redis", "memcached"], version = "^0.11.1" }
aiosmtplib = "^1.1.6"
pyotp = "^2.6.0"
msgpack-python = "^0.5.6"
python-dateutil = "^2.8.2"
orjson = "^3.6.7"
fastapi = { extras = ["all"], version = "^0.75.0" }
httpx = {extras = ["http2"], version = "^0.22.0"}


[tool.poetry.dev-dependencies]
pytest-asyncio = "^0.18.1"
pytest-xdist = "^2.5.0"
pytest-ordering = "^0.6"


[tool.aerich]
tortoise_orm = "settings.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
